feature: refactoring(add JSDOC and move methods)
