+++ src/components/pages/music-player/MusicPlayer.tsx
@@ -10,6 +10,7 @@ import { Icon } from "~/components/common/icon/Icon";
 import { MusicTable } from "~/components/common/table/MusicTable";
 
 import useFile from "~/hooks/useFile";
+import useMusic from "~/hooks/useMusic";
 import useMusicPlayer from "~/hooks/useMusicPlayer";
 
 import { formatSecondsToMMSS } from "~/utils/format";
@@ -18,6 +19,8 @@ import classes from "~/components/pages/music-player/MusicPlayer.module.scss";
 
 export const MusicPlayer: FC = () => {
   const { fileRef, onClickInputFileList } = useFile();
+  // FIXME 一旦直接取得。できればuseMusicPlayerで取得したい。
+  const { getAudioArrayBuffer, loadedAudioMetadata } = useMusic();
   const {
     currentMusic,
     currentMusicList,
@@ -28,9 +31,7 @@ export const MusicPlayer: FC = () => {
     backForward,
     clear,
     forward,
-    getAudioArrayBuffer,
     getMusicURL,
-    loadedAudioMetadata,
     pause,
     play,
     setIsLoading,
